\name{pkg9-package}
\alias{pckg9-package}
\alias{pckg9}
\alias{pack}
\docType{package}
\title{
\packageTitle{pkg9}
}
\description{
\packageDescription{pkg9}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pkg9}
\packageIndices{pkg9}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pkg9}

Maintainer: \packageMaintainer{pkg9}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
myMean(1:50)
}
