% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg9-package.R
\docType{package}
\name{pkg9}
\alias{pkg9}
\alias{pckg9}
\alias{package9}
\alias{pck}
\alias{plk}
\alias{pkage9}
\title{Help page title. Probably the package name and tagline.}
\description{
A description of what the package does, why you might want to use it, #' which functions to look at first, and anything else that the user
really, absolutely, must look at because you've created it and it is #' astonishing.
}
\author{
h4ck3r
}
\keyword{package}
